fix: cartCart 함수를 Store외부의 유틸함수 관리하는 폴더 내부로 이동해서 관리
